# docs/swagger.yaml
definitions:
...
  domain.BindingEntranceZone:
    properties:
      created_at:
        type: string
      creator:
        type: string
      direction:
        type: string
      entrance_id:
        type: string
      entrance_name:
        type: string
      kind_zone:
        type: string
      modified_at:
        type: string
      modifier:
        type: string
      options:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
      zone_id:
        type: string
    type: object
...
  domain.BindingsEntranceZone:
    items:
      $ref: '#/definitions/domain.BindingEntranceZone'
    type: array
...
  domain.MallSensor:
    properties:
      created_at:
        type: string
      creator:
        type: string
      device_id:
        type: string
      external_id:
        type: string
      is_active:
        type: boolean
      kind:
        type: string
      layout_id:
        type: string
      modified_at:
        type: string
      modifier:
        type: string
      notes:
        type: string
      options:
        type: string
      sensor_id:
        type: string
      title:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
  domain.MallSensors:
    items:
      $ref: '#/definitions/domain.MallSensor'
    type: array
  domain.MallZone:
    properties:
      area:
        type: number
      created_at:
        type: string
      creator:
        type: string
      entrances:
        $ref: '#/definitions/domain.BindingsEntranceZone'
        type: object
      is_active:
        type: boolean
      kind:
        type: string
      layout_id:
        type: string
      modified_at:
        type: string
      modifier:
        type: string
      notes:
        type: string
      options:
        type: string
      parent_id:
        type: string
      sensors:
        $ref: '#/definitions/domain.MallSensors'
        type: object
      title:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
      zone_id:
        type: string
    type: object
  domain.MallZones:
    items:
      $ref: '#/definitions/domain.MallZone'
    type: array
...
  domain.Renter:
    properties:
      category_id:
        type: string
      contract:
        type: string
      created_at:
        type: string
      creator:
        type: string
      is_active:
        type: boolean
      layout_id:
        type: string
      modified_at:
        type: string
      modifier:
        type: string
      notes:
        type: string
      options:
        type: string
      price_segment_id:
        type: string
      renter_id:
        type: string
      time_close:
        type: string
      time_open:
        type: string
      title:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
      zones:
        $ref: '#/definitions/domain.MallZones'
        type: object
    type: object
...
  domain.Renters:
    items:
      $ref: '#/definitions/domain.Renter'
    type: array
...
  infra.ErrResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
  infra.HTTPError:
    properties:
      message:
        type: object
    type: object
...
  infra.RentersResponse:
    properties:
      data:
        $ref: '#/definitions/domain.Renters'
        type: object
      result_set:
        $ref: '#/definitions/infra.ResultSet'
        type: object
    type: object
...
  infra.ResultSet:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact:
    email: 1020@watcom.ru
    name: API Support
    url: https://helpdesk.watcom.ru
  description: |-
    This is a general service for interacting with layout configuration.
    Общий сервис для взаимодействия с конфигурацией схем размещения.
  license: {}
  title: LayoutConfig API
  version: 2.0/1.1.21
paths:
...
  /v2/malls/renters:
    get:
      description: |-
        get slice of renters with loc, date, layout_id, category_id, price_segment_id, contract, offset, limit, fields parameters
        layout_id (recommended parameter), if not pass datasource may be not correct
        fields - comma separated values of field renter_id,title,layout_id,price_segment_id...
      parameters:
      - description: uuid format, default=*
        in: query
        name: layout_id
        required: true
        type: string
      - description: default=*
        in: query
        name: categor_id
        type: string
      - description: default=*
        in: query
        name: price_segment_id
        type: string
      - description: default=*
        in: query
        name: contract
        type: string
      - description: location, default=ru
        in: query
        name: loc
        type: string
      - description: default=0
        in: query
        name: offset
        type: integer
      - description: default=20
        in: query
        name: limit
        type: integer
      - description: renter_id,title,layout_id... default=all
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/infra.RentersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get all renters in the mall schema
      tags:
      - malls/renters
  /v2/malls/renters/{renter_id}:
    get:
      description: |-
        get renter with loc, date, renter_id, fields parameters
        layout_id (recommended parameter), if not pass datasource may be not correct
        fields - comma separated values of field renter_id,title,layout_id,price_segment_id...
        include - comma separated list of entities, embedded in current, for renters it can be only zones
      parameters:
      - description: uuid format, default=*
        in: query
        name: layout_id
        required: true
        type: string
      - description: uuid/digits format
        in: path
        name: renter_id
        required: true
        type: string
      - description: default=ru
        in: query
        name: loc
        type: string
      - description: ISO8601 YYYY-MM-DD date
        in: query
        name: date
        type: string
      - description: renter_id,title,layout_id... default=all
        in: query
        name: fields
        type: string
      - description: zones default=none
        in: query
        name: include
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Renter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get specified renter in the mall schema
      tags:
      - malls/renters
...
swagger: "2.0"